Add metaclass for main widget, improve logo, add title
